Fix clang-tidy errors.

+++ src/satellite_stream_transporter.cpp
@@ -1,15 +1,15 @@
 #include "satellite_stream_transporter.h"
 
 #include <arpa/inet.h>
-#include <errno.h>
 #include <netdb.h>
 #include <netinet/in.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include <cerrno>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
 
 #include <stdexcept>
 
@@ -21,7 +21,8 @@ SatelliteStreamTransporter::SatelliteStreamTransporter(
     const LightStepTracerOptions& options) {
   sockaddr_in satellite_address = {};
   satellite_address.sin_family = AF_INET;
-  satellite_address.sin_port = htons(options.collector_port);
+  satellite_address.sin_port =
+      htons(static_cast<uint16_t>(options.collector_port));
 
   auto rcode = inet_pton(AF_INET, options.collector_host.c_str(),
                          &satellite_address.sin_addr);
